Pull up item matching
Both remove and update need to match existing items against an item
spec. Pulled the common code into the superclass
